refactoring: move counter stuff to store/counter
