+++ frontend/railwayspringapp/src/app/app.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
-import { MyCounterComponent } from './my-counter/my-counter.component';
+import { MyCounterComponent } from './store/my-counter/my-counter.component';
 
 @Component({
   selector: 'app-root',
