+++ frontend/railwayspringapp/src/app/app.config.ts
@@ -6,7 +6,7 @@ import { provideStore } from '@ngrx/store';
 import { provideEffects } from '@ngrx/effects';
 import { provideStoreDevtools } from '@ngrx/store-devtools';
 
-import { counterReducer } from './counter.reducer';
+import { counterReducer } from './store/my-counter/counter.reducer';
 
 export const appConfig: ApplicationConfig = {
   providers: [
